Move constant speed test to Finland test set, add serviceway  nothrough test
