+++ application/src/test/java/org/opentripplanner/osm/tagmapping/FinlandMapperTest.java
@@ -1,10 +1,12 @@
 package org.opentripplanner.osm.tagmapping;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.opentripplanner.street.model.StreetTraversalPermission.NONE;
 import static org.opentripplanner.street.model.StreetTraversalPermission.PEDESTRIAN;
 import static org.opentripplanner.street.model.StreetTraversalPermission.PEDESTRIAN_AND_BICYCLE;
 
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.opentripplanner.osm.model.OsmWay;
 import org.opentripplanner.osm.model.OsmWithTags;
@@ -13,12 +15,15 @@ import org.opentripplanner.osm.wayproperty.WayPropertySet;
 
 public class FinlandMapperTest {
 
-  static WayPropertySet wps = new WayPropertySet();
+  private WayPropertySet wps;
+  private OsmTagMapper mapper;
   static float epsilon = 0.01f;
 
-  static {
-    var source = new FinlandMapper();
-    source.populateProperties(wps);
+  @BeforeEach
+  public void setup() {
+    this.wps = new WayPropertySet();
+    this.mapper = new FinlandMapper();
+    this.mapper.populateProperties(this.wps);
   }
 
   /**
@@ -219,5 +224,27 @@ public class FinlandMapperTest {
     way.addTag("bicycle", "yes");
     wayData = wps.getDataForWay(way);
     assertEquals(wayData.getPermission(), PEDESTRIAN_AND_BICYCLE);
+  }
+
+  @Test
+  public void constantSpeedCarRouting() {
+    OsmTagMapper osmTagMapper = new ConstantSpeedFinlandMapper(20f);
+
+    var slowWay = new OsmWithTags();
+    slowWay.addTag("highway", "residential");
+    assertEquals(20f, osmTagMapper.getCarSpeedForWay(slowWay, true));
+
+    var fastWay = new OsmWithTags();
+    fastWay.addTag("highway", "motorway");
+    fastWay.addTag("maxspeed", "120 kmph");
+    assertEquals(20f, osmTagMapper.getCarSpeedForWay(fastWay, true));
+  }
+
+  @Test
+  public void serviceNoThroughTraffic() {
+    var way = new OsmWay();
+    way.addTag("highway", "residential");
+    way.addTag("service", "driveway");
+    assertTrue(mapper.isMotorVehicleThroughTrafficExplicitlyDisallowed(way));
   }
 }
